add python3 suppport based on 2to3 script

+++ fields/integer.py
@@ -60,7 +60,7 @@ class Integer(Field):
                        height, maxChars, colspan, master, masterValue, focus,
                        historized, mapping, label, sdefault, scolspan, swidth,
                        sheight, persist, view, xml)
-        self.pythonType = long
+        self.pythonType = int
 
     def validateValue(self, obj, value):
         try:
